+++ src/commonTest/kotlin/com/bkahlert/kommons/test/RootCauseKtTest.kt
@@ -1,8 +1,11 @@
 package com.bkahlert.kommons.test
 
+import io.kotest.matchers.should
+import io.kotest.matchers.shouldBe
+import io.kotest.matchers.types.shouldBeInstanceOf
 import kotlin.test.Test
 
-class ExceptionsMatchersKtTest {
+class RootCauseKtTest {
 
     @Test fun root_cause() = testAll(
         IllegalArgumentException("error message"),
@@ -11,6 +14,10 @@ class ExceptionsMatchersKtTest {
         Error(RuntimeException(IllegalStateException(IllegalArgumentException("error message")))),
         RuntimeException(Error(RuntimeException(IllegalStateException(IllegalArgumentException("error message"))))),
     ) { ex ->
+        ex.rootCause should {
+            it.shouldBeInstanceOf<IllegalArgumentException>()
+            it.message shouldBe "error message"
+        }
         ex.shouldHaveRootCauseInstanceOf<IllegalArgumentException>()
         ex.shouldHaveRootCauseMessage("error message")
     }